introduce variants for popularity

+++ modules/TripletSampler.py
@@ -12,11 +12,11 @@ class TripletSampler:
 
     def set_selection_list(self, most_popular_items, step):
         if most_popular_items[0] == 1:
-            s = step * 0.1 * len(most_popular_items[1])
-            e = (step + 1) * 0.1 * len(most_popular_items[1])
+            s = int(step * 0.1 * len(most_popular_items[1]))
+            e = int((step + 1) * 0.1 * len(most_popular_items[1]))
         elif most_popular_items[0] == 2:
             s = 0
-            e = (step + 1) * 0.1 * len(most_popular_items[1])
+            e = int((step + 1) * 0.1 * len(most_popular_items[1]))
         self.selection_list = list(set(self.train_user_list) & set(most_popular_items[1][s:e]))
 
     def sample_user_triples(self):
